Tutorial Scene
Updated the UI of the infoScene to set up a tutorial on how to play the game. Created a simple dialogue system for the tutorial level, introducing the NPC character and explaining how to drag mixtures into boxes. Also added buttons that the player can use to return to the Home screen or start the game.